backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.17
django-extensions==3.2.3
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
importlib_metadata==8.5.0
jwcrypto==1.5.6
Markdown==3.7
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
